ToDoアプリのエラーハンドリングをするPart4

+++ src/Laravel9TaskList/app/Http/Controllers/FolderController.php
@@ -14,6 +14,9 @@ use App\Http\Requests\EditFolder;
 // Authクラスをインポートする
 use Illuminate\Support\Facades\Auth;
 
+use Illuminate\Support\Facades\DB;
+
+use Illuminate\Support\Facades\Log;
 
 class FolderController extends Controller
 {
@@ -28,9 +31,20 @@ class FolderController extends Controller
         /* フォルダの新規作成ページを呼び出す */
         // view('遷移先のbladeファイル名');
         // return view('folders/create');
-        $folders = Auth::user()->folders;
+        // $folders = Auth::user()->folders;
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            $user->folders;
+
+            // return view('folders/create', compact('folders'));
+            return view('folders/create');
+
+        } catch (\Throwable $e) {
+            Log::error('Error FolderController in showCreateForm: ' . $e->getMessage());
+        }
         
-        return view('folders/create', compact('folders'));
+
 
     }
 
@@ -46,31 +60,35 @@ class FolderController extends Controller
     //  *  @param Request $request （リクエストクラスの$request）
     public function create(CreateFolder $request)
     {
-        /* 新規作成のフォルダー名（タイトル）をDBに書き込む処理 */
-        // フォルダモデルのインスタンスを作成する
-        $folder = new Folder();
-        // タイトルに入力値を代入する
-        $folder->title = $request->title;
-        // インスタンスの状態をデータベースに書き込む
-        // $folder->save();
-        
-                // （ログイン）ユーザーに紐づけて保存する
-                // Auth::user()->folders()->save($folder);
-                /** @var App\Models\User **/
-                $user = Auth::user();
-                $user->folders()->save($folder);
+        try {
+            /* 新規作成のフォルダー名（タイトル）をDBに書き込む処理 */
+            // フォルダモデルのインスタンスを作成する
+            $folder = new Folder();
+            // タイトルに入力値を代入する
+            $folder->title = $request->title;
+            // インスタンスの状態をデータベースに書き込む
+            // $folder->save();
+            
+                    // （ログイン）ユーザーに紐づけて保存する
+                    // Auth::user()->folders()->save($folder);
+                    /** @var App\Models\User **/
+                    $user = Auth::user();
+                    $user->folders()->save($folder);
 
 
-        /* タスク一覧ページにリダイレクトする */
-        // リダイレクト：別URLへの転送（リクエストされたURLとは別のURLに直ちに再リクエストさせます）
-        // route('遷移先のbladeファイル名', [連想配列：渡したい変数についての情報]);
-        // 連想配列：['キー（テンプレート側で参照する際の変数名）' => '渡したい変数']
-        // redirect():リダイレクトを実施する関数
-        // route():ルートPathを指定する関数
-        return redirect()->route('tasks.index', [
-            // 'id' => $folder->id,
-            'folder' => $folder->id,
-        ]);
+            /* タスク一覧ページにリダイレクトする */
+            // リダイレクト：別URLへの転送（リクエストされたURLとは別のURLに直ちに再リクエストさせます）
+            // route('遷移先のbladeファイル名', [連想配列：渡したい変数についての情報]);
+            // 連想配列：['キー（テンプレート側で参照する際の変数名）' => '渡したい変数']
+            // redirect():リダイレクトを実施する関数
+            // route():ルートPathを指定する関数
+            return redirect()->route('tasks.index', [
+                // 'id' => $folder->id,
+                'folder' => $folder->id,
+            ]);
+        } catch (\Exception $e) {
+            Log::error('Error FolderController in create: ' . $e->getMessage());
+        }
     }
     /**
      *  【フォルダ編集ページの表示機能】
@@ -84,16 +102,20 @@ class FolderController extends Controller
     // public function showEditForm(int $id)
     public function showEditForm(Folder $folder)
     {
-               /** @var App\Models\User **/
-               $user = Auth::user();
-        // $folder = Folder::find($id);
-        // $folder = $user->folders()->findOrFail($id);
-        $folder = $user->folders()->findOrFail($folder->id);
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            // $folder = Folder::find($id);
+            // $folder = $user->folders()->findOrFail($id);
+            $folder = $user->folders()->findOrFail($folder->id);
 
-        return view('folders/edit', [
-            'folder_id' => $folder->id,
-            'folder_title' => $folder->title,
-        ]);
+            return view('folders/edit', [
+                'folder_id' => $folder->id,
+                'folder_title' => $folder->title,
+            ]);
+        } catch (\Throwable $e) {
+            Log::error('Error FolderController in showEditForm: ' . $e->getMessage());
+        }
     }
     /**
      *  【フォルダの編集機能】
@@ -107,19 +129,32 @@ class FolderController extends Controller
     // public function edit(int $id, EditFolder $request)
     public function edit(Folder $folder, EditFolder $request)
     {
-        /** @var App\Models\User **/
-        $user = Auth::user();
-        //   $folder = $user->folders()->findOrFail($id);
-        $folder = $user->folders()->findOrFail($folder->id);
-        // $folder = Folder::find($id);
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            //   $folder = $user->folders()->findOrFail($id);
+            $folder = $user->folders()->findOrFail($folder->id);
+            // $folder = Folder::find($id);
 
-        $folder->title = $request->title;
-        $folder->save();
+            // $folder->title = $request->title;
+            // $folder->save();
 
-        return redirect()->route('tasks.index', [
-            // 'id' => $folder->id,
-            'folder' => $folder->id,
-        ]);
+            $folder = DB::transaction(function () use ($folder) {
+                if($folder) throw new \Exception('500');
+                $folder->tasks()->delete();
+                $folder->delete();
+                return $folder;
+            });
+
+            $folder = Folder::first();
+
+            return redirect()->route('tasks.index', [
+                // 'id' => $folder->id,
+                'folder' => $folder->id,
+            ]);
+        } catch (\Throwable $e) {
+            Log::error('Error FolderController in delete: ' . $e->getMessage());
+        }
     }
 
 
@@ -160,24 +195,31 @@ class FolderController extends Controller
     // public function delete(int $id)
     public function delete(Folder $folder)
     {
-        
-         /** @var App\Models\User **/
-         $user = Auth::user();
-         
-        // $folder = Folder::find($id);
-        //  $folder = $user->folders()->findOrFail($id);
-        $folder = $user->folders()->findOrFail($folder->id);
- 
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            
+            // $folder = Folder::find($id);
+            //  $folder = $user->folders()->findOrFail($id);
+            $folder = $user->folders()->findOrFail($folder->id);
 
-        $folder->tasks()->delete();
-        $folder->delete();
+            $folder = DB::transaction(function () use ($folder) {
+                if($folder) throw new \Exception('500');
 
-        $folder = Folder::first();
+                        $folder->tasks()->delete();
+                        $folder->delete();
+                        return $folder;
+            });
 
-        return redirect()->route('tasks.index', [
-            // 'id' => $folder->id
-            'folder' => $folder->id
-        ]);
+            $folder = Folder::first();
+
+            return redirect()->route('tasks.index', [
+                // 'id' => $folder->id
+                'folder' => $folder->id
+            ]);
+        } catch (\Throwable $e) {
+            Log::error('Error FolderController in delete: ' . $e->getMessage());
+        }
     }
